CommonLib, ClientLib: 클라/서버 공용으로 사용하는 핵심 전투 코드에서 사용되는 float 제거

+++ Assets/Script/ClientLib/Network/App/TestClientApp.cs
@@ -86,7 +86,7 @@ namespace Script.ClientLib.Network.App
             return true;
         }
         
-        public async Task<List<Tuple<uint, float>>> RequestVerifyStageBattle(List<ushort> updateIntervals)
+        public async Task<List<Tuple<uint, uint>>> RequestVerifyStageBattle(List<ushort> updateIntervals)
         {
             if (!_ctx.IsInitialized)
             {
