+++ Assets/Script/CommonLib/Map/IBattleMapContext.cs
@@ -7,7 +7,7 @@ namespace Script.CommonLib.Map
         public void OnEntityPositionChanged(uint entityId, FixedPos pos);
         public void OnEntityDirectionChanged(uint entityId, FixedDir dir);
         public void OnEntityGetDamage(uint entityId, uint damage);
-        public IEntityContext TryGetNearestEnemy(uint entityId, float maxDistance);
+        public IEntityContext TryGetNearestEnemy(uint entityId, long maxDistance);
         public void FindWaypoints(GridPos start, GridPos goal, List<GridPos> resultWaypoints);
         public uint ElapsedMs { get; }
         public void RequestAttack(uint attackerId, uint targetEntityId);
